todo-app/backend-todo/requirements.txt
fastapi>=0.128.0
sqlmodel>=0.0.22
python-dotenv>=1.2.1
uvicorn>=0.40.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0